+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Block meshes now end in `_0` suffix to make auto-incrementing cleaner in map placement.
 - `MultiplayerTest` now monitors the status of the spawned processes and updates state accordingly.
+- `ANSamplesPawn` now replicates movement.
 
 ### Removed